Fix undefined behavior in C MUL by using int64_t cast
Signed overflow in C is undefined behavior. MUL must wrap modulo 2^32.
Cast operands to int64_t before multiplication, then truncate to int32_t
to guarantee correct wrapping semantics under optimization.

+++ c/main.c
@@ -142,7 +142,7 @@ void tick() {
 	case 0b01100000:// add/sub/mul
 		registers[register_destination] = (
 			instruction & (1 << 25) // mul?
-			?	registers[register_source1] * registers[register_source2]
+			?	(int32_t) ((int64_t) registers[register_source1] * (int64_t) registers[register_source2])
 			: instruction >> 30 // sub?
 			?	registers[register_source1] - registers[register_source2]
 			:	registers[register_source1] + registers[register_source2]
